requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
# Pin NumPy below 2.0 due to compiled extension compatibility
numpy>=1.24,<2.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Tracking
tqdm>=4.65.0

# Utilities
opencv-python>=4.8.0
albumentations>=1.3.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.25.0

# Metrics and Logging
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0.0